clean code need to test amp

+++ src/utils/fourier.py
@@ -63,7 +63,7 @@ def real_imag_to_complex_freq(x: torch.Tensor):
     x_im = x[:, n_real:, :]
 
     # Create imaginary tensor
-    zero_padding = torch.zeros(size=(x.size(0), 1, x.size(2)), device=x.device)
+    zero_padding = torch.zeros(size=(x.size(0), 1, x.size(2)), device=x.device, dtype=x.dtype)
     x_im = torch.cat((zero_padding, x_im), dim=1)
 
     # If number of time steps is even, put the null imaginary part
